DinoBot/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
idna==2.9
MouseInfo==0.1.3
Pillow==7.1.2
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
requests==2.23.0
selenium==3.141.0
urllib3==1.25.9
webdriver-manager==2.5.2
